third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/the-button-element/button-validation-expected.txt

This is a testharness.js-based test.
PASS missing type attribute
PASS submit type attribute
PASS reset type attribute
PASS button type attribute
FAIL menu type attribute assert_equals: expected "menu" but got "submit"
PASS invalid type attribute
Harness: the test ran to completion.
